requirements.txt
markdown>=3.5
Jinja2>=3.1
weasyprint>=60.0
pyyaml>=6.0
